// src/types/Cliente.ts
export type ClienteProps = {
  nomeCliente: string;
  dataNascimento: string;
  apelido: string;
  cpfcnpj: string;
  rgie?: string;
  nomeFantasia?: string;
  razaoSocial?: string;
  email: string;
  senha: string;
  cep: string;
  celular1: string;
  celular2: string;
  ativo: boolean;
  newsletter: boolean;
};